# Conflicts:
#	src/test/java/seedu/address/model/predicate/AbsentFromTutorialPredicateTest.java

+++ src/main/java/seedu/address/logic/parser/DeleteCommandParser.java
@@ -1,10 +1,13 @@
 package seedu.address.logic.parser;
 
 import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
+import static seedu.address.logic.parser.CliSyntax.PREFIX_TUTORIALGROUP;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.DeleteCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
+import seedu.address.model.predicate.ContainsTagPredicate;
+import seedu.address.model.tag.Tag;
 
 /**
  * Parses input arguments and creates a new DeleteCommand object
@@ -16,14 +19,34 @@ public class DeleteCommandParser implements Parser<DeleteCommand> {
      * and returns a DeleteCommand object for execution.
      * @throws ParseException if the user input does not conform the expected format
      */
-    public DeleteCommand parse(String args) throws ParseException {
-        try {
-            Index index = ParserUtil.parseIndex(args);
-            return new DeleteCommand(index);
-        } catch (ParseException pe) {
-            throw new ParseException(
-                    String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_USAGE), pe);
-        }
-    }
 
+    public DeleteCommand parse(String args) throws ParseException {
+
+        String trimmedArgs = args.trim();
+        if (!trimmedArgs.startsWith("all")) {
+            try {
+                Index index = ParserUtil.parseIndex(args);
+                return new DeleteCommand(index);
+            } catch (ParseException pe) {
+                throw new ParseException(
+                        String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_USAGE), pe);
+            }
+        }
+
+        ArgumentMultimap argMultimap = ArgumentTokenizer.tokenize(args, PREFIX_TUTORIALGROUP);
+
+        if (trimmedArgs.equals("all")) {
+            Tag tag = new Tag("PLACEHOLDER");
+            return new DeleteCommand(tag, new ContainsTagPredicate(tag));
+        }
+
+        if (!argMultimap.getValue(PREFIX_TUTORIALGROUP).isPresent()) {
+            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_USAGE));
+        }
+
+        String tagName = argMultimap.getValue(PREFIX_TUTORIALGROUP).get();
+        Tag tag = ParserUtil.parseTag(tagName);
+
+        return new DeleteCommand(tag, new ContainsTagPredicate(tag));
+    }
 }
